Use float

+++ model.py
@@ -21,11 +21,11 @@ class ChEBIRecNN(nn.Module):
 
         self.norm = torch.nn.LayerNorm(self.length)
 
-        self.NN_single_node = nn.Sequential(nn.Linear(self.atom_enc, self.length), nn.ReLU(), nn.Linear(self.length, self.length)).double()
-        self.merge = nn.Sequential(nn.Linear(2*self.length, self.length), nn.ReLU(), nn.Linear(self.length, self.length)).double()
-        self.register_parameter("attention_weight", torch.nn.Parameter(torch.rand(self.length,1, requires_grad=True).double()))
-        self.register_parameter("dag_weight", torch.nn.Parameter(torch.rand(self.length,1, requires_grad=True).double()))
-        self.final = nn.Linear(self.length, self.num_of_classes).double()
+        self.NN_single_node = nn.Sequential(nn.Linear(self.atom_enc, self.length), nn.ReLU(), nn.Linear(self.length, self.length))
+        self.merge = nn.Sequential(nn.Linear(2*self.length, self.length), nn.ReLU(), nn.Linear(self.length, self.length))
+        self.register_parameter("attention_weight", torch.nn.Parameter(torch.rand(self.length,1, requires_grad=True)))
+        self.register_parameter("dag_weight", torch.nn.Parameter(torch.rand(self.length,1, requires_grad=True)))
+        self.final = nn.Linear(self.length, self.num_of_classes)
 
     def forward(self, molecule: Molecule):
         final_outputs = None
